requirements.txt
Flask
gunicorn
line-bot-sdk
imgurpython
Jinja2
wolframalpha